refactor, make voting_test as integration test
