+++ rust/catalyst-voting/src/crypto/mod.rs
@@ -2,5 +2,3 @@
 
 pub(crate) mod elgamal;
 pub(crate) mod group;
-
-pub use elgamal::{PublicKey, SecretKey};
